move prover tests to integration tests.
